Added ICA-based artefact rejection and source reconstruction. Added standard level of filtering.
Working with the motor movement dataset, but would need to get adapted for the collected dataset to take in the csv input. Since we don't have EOG in the collected we'll need to pick a channel that's close to the eyes as an analog and update the channel names for that.